fix wrong folder structure
