+++ src/main/java/com/plumbly/models/Address.java
@@ -1,4 +1,4 @@
-package com.plumbly.plumbly.models;
+package com.plumbly.models;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
