+++ src/main/java/com/plumbly/models/Profession.java
@@ -1,24 +1,21 @@
-package com.plumbly.plumbly.models;
+package com.plumbly.models;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 @Document
-public class Country {
+public class Profession {
     @Id
     private String id;
     private String name;
-    private String isoCode;
-    private String phonePrefix;
+    private boolean requiresCertificate;
 
     @Override
     public String toString() {
-        return "Country{" +
+        return "Profession{" +
                 "id='" + id + '\'' +
                 ", name='" + name + '\'' +
-                ", isoCode='" + isoCode + '\'' +
-                ", phonePrefix='" + phonePrefix + '\'' +
+                ", requiresCertificate=" + requiresCertificate +
                 '}';
     }
 }
-
