+++ src/main/java/com/plumbly/repositories/AddressRepository.java
@@ -1,8 +1,8 @@
-package com.plumbly.plumbly.repositories;
+package com.plumbly.repositories;
 
 import org.springframework.data.repository.CrudRepository;
 
-import com.plumbly.plumbly.models.Address;
+import com.plumbly.models.Address;
 
 public interface AddressRepository extends CrudRepository<Address, String>{
