+++ src/main/java/com/plumbly/repositories/CompanyRepository.java
@@ -1,8 +1,8 @@
-package com.plumbly.plumbly.repositories;
+package com.plumbly.repositories;
 
 import org.springframework.data.repository.CrudRepository;
 
-import com.plumbly.plumbly.models.Company;
+import com.plumbly.models.Company;
 
 public interface CompanyRepository extends CrudRepository<Company, String>{
