+++ src/main/java/com/plumbly/repositories/HouseholderRepository.java
@@ -1,8 +1,8 @@
-package com.plumbly.plumbly.repositories;
+package com.plumbly.repositories;
 
 import org.springframework.data.repository.CrudRepository;
 
-import com.plumbly.plumbly.models.Householder;
+import com.plumbly.models.Householder;
 
 public interface HouseholderRepository extends CrudRepository<Householder, String>{
