+++ src/main/java/com/plumbly/repositories/TechnicianRepository.java
@@ -1,7 +1,7 @@
-package com.plumbly.plumbly.repositories;
+package com.plumbly.repositories;
 
 import org.springframework.data.repository.CrudRepository;
-import com.plumbly.plumbly.models.Technician;
+import com.plumbly.models.Technician;
 
 public interface TechnicianRepository extends CrudRepository<Technician, String>{
