+++ src/main/java/com/plumbly/repositories/UserRepository.java
@@ -1,8 +1,8 @@
-package com.plumbly.plumbly.repositories;
+package com.plumbly.repositories;
 
 import org.springframework.data.repository.CrudRepository;
 
-import com.plumbly.plumbly.models.User;
+import com.plumbly.models.User;
 
 public interface UserRepository extends CrudRepository<User, String> {
