+++ src/test/java/com/plumbly/PlumblyApplicationTests.java
@@ -1,4 +1,4 @@
-package com.plumbly.plumbly;
+package com.plumbly;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
